refactor : 패키지 분리
- Domain, presentation 우선 분리
- MovingStrategy 의 메서드명 shouldMove() 로 변경
- ErrorMessage 는 global 패키지로 관리
