+++ src/main/java/racingcar/presentation/InputView.java
@@ -2,7 +2,7 @@ package racingcar.presentation;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.NoSuchElementException;
-import racingcar.ErrorMessage;
+import racingcar.global.ErrorMessage;
 
 public class InputView {
     private static final String INPUT_CARS_NAMES = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
