+++ src/test/java/racingcar/presentation/InputValidatorTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
-import racingcar.ErrorMessage;
+import racingcar.global.ErrorMessage;
 
 class InputValidatorTest {
