+++ src/test/java/racingcar/presentation/InputViewTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.ErrorMessage;
+import racingcar.global.ErrorMessage;
 
 class InputViewTest {
     private InputView inputView;
